+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Migrations/IdempotentApplyMigrations.sql
@@ -1540,3 +1540,29 @@ BEGIN
 END $EF$;
 COMMIT;
 
+START TRANSACTION;
+
+
+DO $EF$
+BEGIN
+    IF NOT EXISTS(SELECT 1 FROM "__EFMigrationsHistory" WHERE "MigrationId" = '20250109184346_FixResourceHoldersIndex') THEN
+    DROP INDEX "IX_resource_holders_entity_id_resource_entity_id_balance";
+    END IF;
+END $EF$;
+
+DO $EF$
+BEGIN
+    IF NOT EXISTS(SELECT 1 FROM "__EFMigrationsHistory" WHERE "MigrationId" = '20250109184346_FixResourceHoldersIndex') THEN
+    CREATE INDEX "IX_resource_holders_resource_entity_id_balance_id" ON resource_holders (resource_entity_id, balance, id);
+    END IF;
+END $EF$;
+
+DO $EF$
+BEGIN
+    IF NOT EXISTS(SELECT 1 FROM "__EFMigrationsHistory" WHERE "MigrationId" = '20250109184346_FixResourceHoldersIndex') THEN
+    INSERT INTO "__EFMigrationsHistory" ("MigrationId", "ProductVersion")
+    VALUES ('20250109184346_FixResourceHoldersIndex', '8.0.2');
+    END IF;
+END $EF$;
+COMMIT;
+